Fixed problem with grouping interval (double => long, for correct division), added not finished EndToEnd tests.

+++ FlowUpEntryTest.Abstractions/IEndpointRequestRegister.cs
@@ -7,6 +7,6 @@
         public int[] GetRequestsCountsWithinInterval(string endpointName,
             DateTime from,
             DateTime to,
-            double intervalMinutes);
+            long intervalMinutes);
     }
 }
